Support newer "AWS Control Tower Account Factory" versions
Since the last Control Tower update, the Service Catalog product
contains a path ID that needs to be provided when provisioning an
account using the Account Factory.

We attempt to look it up, and if we find it use it, or skip it if not
found to support older versions. If we find more than one path ID then
error out and ask the user to supply the path ID by hand using the
`provisioned_product_path_id` field.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>